Fix remaining warnings on tal plugins

+++ libs/juce/source/modules/juce_audio_plugin_client/LV2/juce_LV2_Wrapper.cpp
@@ -1505,7 +1505,7 @@ public:
 
                 while (i.getNextEvent (midiEventData, midiEventSize, midiEventPosition))
                 {
-                    jassert (midiEventPosition >= 0 && midiEventPosition < sampleCount);
+                    jassert (midiEventPosition >= 0 && midiEventPosition < (int)sampleCount);
 
                     if (sizeof(LV2_Atom_Event) + midiEventSize > capacity - offset)
                         break;
